support x86_64 build

+++ linuxdoom-1.10/r_draw.c
@@ -38,6 +38,8 @@
 #include "v_video.h"
 
 // State.
+#include <stdint.h>
+
 #include "doomstat.h"
 
 
@@ -459,7 +461,7 @@ void R_InitTranslationTables (void)
     int		i;
 	
     translationtables = Z_Malloc (256*3+255, PU_STATIC, 0);
-    translationtables = (byte *)(( (int)translationtables + 255 )& ~255);
+    translationtables = (byte *)(( (intptr_t)translationtables + 255 )& ~255);
     
     // translate just the 16 green colors
     for (i=0 ; i<256 ; i++)
